Now doesn't totally fail in drawing clusters for large process counts - However, what to do when the majority of processes are not doing anything at all?

+++ gnome.h
@@ -100,14 +100,15 @@ protected:
     void drawGnomeQtTopProcesses(QPainter * painter, QRect extents,
                                  int blockwidth, int barwidth);
     void drawGnomeQtClusterBranch(QPainter * painter, QRect current, PartitionCluster * pc,
-                                  int blockheight, int blockwidth, int barheight, int barwidth);
-    void drawGnomeQtClusterLeaf(QPainter * painter, QRect startxy, QList<Event *> * elist, int blockwidth, int startStep);
-    void drawGnomeQtInterMessages(QPainter * painter, int blockwidth, int startStep, int startx);
+                                  float blockheight, int blockwidth, int barheight, int barwidth);
+    void drawGnomeQtClusterLeaf(QPainter * painter, QRect startxy, QList<Event *> * elist,
+                                int blockwidth, int startStep); // Doesn't use blockheight
+    void drawGnomeQtInterMessages(QPainter * painter, int blockwidth, int startStep, int startx); // Doesn't use blockheight
     virtual void drawGnomeQtClusterEnd(QPainter * painter, QRect clusterRect, PartitionCluster * pc,
                                     int barwidth, int barheight, int blockwidth, int blockheight,
-                                    int startStep);
+                                    int startStep); // At this point, blockheight will just get recalculated anyway probably
     void drawTreeBranch(QPainter * painter, QRect current, PartitionCluster * pc,
-                        int branch_length, int labelwidth, int blockheight, int leafx);
+                        int branch_length, int labelwidth, float blockheight, int leafx);
     int getTopHeight(QRect extents);
 
     static const int clusterMaxHeight = 76;
